requirements.txt
langchain
langchain-google-genai
langchain-text-splitters
langchain-community
langchain-chroma
langgraph
python-dotenv
pypdf
chromadb
streamlit
